ft: change data type change for chainid

+++ backend/src/models/sponsorship-policy.ts
@@ -112,7 +112,7 @@ export function initializeSponsorshipPolicyModel(sequelize: Sequelize, schema: s
             field: 'IS_APPLICABLE_TO_ALL_NETWORKS'
         },
         enabledChains: {
-            type: DataTypes.ARRAY(DataTypes.INTEGER),
+            type: DataTypes.ARRAY(DataTypes.BIGINT),
             allowNull: true,
             field: 'ENABLED_CHAINS'
         },
